Potential fix for code scanning alert no. 10: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. We can achieve this by casting one of the operands to size_t before performing the multiplication. This will ensure that the multiplication is done using the larger type, preventing overflow.

Specifically, we will cast pv->rows to size_t before performing the multiplication on line 74. This change will ensure that the multiplication is done using the size_t type, which is large enough to hold the result without overflow.

+++ libicu4c-64-2/common/propsvec.cpp
@@ -71,7 +71,7 @@ upvec_open(int32_t columns, UErrorCode *pErrorCode) {
 
     /* set the all-Unicode row and the special-value rows */
     row=pv->v;
-    uprv_memset(row, 0, pv->rows*columns*4);
+    uprv_memset(row, 0, (size_t)pv->rows*columns*4);
     row[0]=0;
     row[1]=0x110000;
     row+=columns;
